+++ docs/whatsnew.md
@@ -7,6 +7,8 @@
 - fix : correct repr of Variable (VT was incorrect)
 - fix : bug reading SNMP variables
 - new property 'isStructure' for Variable object
+- new property 'cpuInfoExtended' for Cpu object
+
 
 ## V1.2.9